+++ lib/TextDevices/ShortcutsDevice.cpp
@@ -24,7 +24,7 @@ namespace TextDevices {
     ShortcutsDevice::dispatch(API* api, Command* command) {
         char pin[4];
         char newCommand[32];
-        uint32_t value;
+        uint16_t value;
 
         if (1 == sscanf(command->original, "read %4s", pin)) {
             snprintf(newCommand, 32, "pin %s read", pin);
@@ -32,13 +32,13 @@ namespace TextDevices {
             return api->dispatch(command);
         }
 
-        if (2 == sscanf(command->original, "write %4s %u", pin, &value)) {
-            snprintf(newCommand, 32, "pin %s write %u", pin, value);
+        if (2 == sscanf(command->original, "write %4s %hu", pin, &value)) {
+            snprintf(newCommand, 32, "pin %s write %hu", pin, value);
             command->body = newCommand;
             return api->dispatch(command);
         }
 
-        if (2 == sscanf(command->original, "pwm %4s %u", pin, &value)) {
+        if (2 == sscanf(command->original, "pwm %4s %hu", pin, &value)) {
             snprintf(newCommand, 32, "pin %s config analog output", pin);
             command->body = newCommand;
             api->dispatch(command);
@@ -46,7 +46,7 @@ namespace TextDevices {
                 // error already reported
                 return true;
             }
-            snprintf(newCommand, 32, "pin %s write %u", pin, value);
+            snprintf(newCommand, 32, "pin %s write %hu", pin, value);
             command->body = newCommand;
             return api->dispatch(command);
         }
